Flatten project
Remove the /sidewalk prefix and incorporate those endpoints directly
into the main file, move db.py into src dir. Only pin the dependencies
directly used, not the dependencies of dependences. Add a default
URL_ROOT.
