+++ .gitignore
@@ -8,6 +8,7 @@ __pycache__/
 
 # Distribution / packaging
 .Python
+ve/
 build/
 develop-eggs/
 dist/
