+++ requirements.txt
@@ -1,28 +1,7 @@
-asgiref==3.4.1
 asyncpg==0.24.0
-attrs==21.2.0
-certifi==2021.5.30
-click==8.0.1
-click-plugins==1.1.1
-cligj==0.7.2
 fastapi==0.68.1
-Fiona==1.8.20
 geopandas==0.9.0
 gunicorn==20.1.0
-h11==0.12.0
-httptools==0.3.0
-munch==2.5.0
-numpy==1.21.2
-pandas==1.3.3
-psycopg2-binary==2.9.1
-pydantic==1.8.2
-pyproj==3.2.1
-python-dateutil==2.8.2
 python-dotenv==0.19.0
-pytz==2021.1
-Shapely==1.7.1
-six==1.16.0
-starlette==0.14.2
-typing-extensions==3.10.0.2
+shapely==1.7.1
 uvicorn==0.15.0
-uvloop==0.16.0
